Updated module 1 readme, renaming lab files
