+++ src/labs/01_data_types_and_variables/mad_lib_solved.rb
